test-python-everi59/solution/requirements.txt
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
Werkzeug==3.0.1
fastapi==0.110.0
pydantic==2.6.3
psycopg2-binary==2.9.9
uvicorn==0.27.1
environs==10.3.0
PyJWT==2.8.0
passlib==1.7.4
